Add youtube resolvers
Added new namespace for youtube tracks and search resolver
