--- src/TobysBot.Music/Search/SpotifyPlaylist.cs
@@ -1,24 +1,0 @@
-using SpotifyAPI.Web;
-
-namespace TobysBot.Music.Search;
-
-public class SpotifyPlaylist : IPlaylist
-{
-    public SpotifyPlaylist(FullAlbum album, IEnumerable<ITrack> tracks)
-    {
-        Tracks = tracks;
-        Url = $"https://open.spotify.com/album/{album.Id}";
-        Title = album.Name;
-    }
-
-    public SpotifyPlaylist(FullPlaylist playlist, IEnumerable<ITrack> tracks)
-    {
-        Tracks = tracks;
-        Url = $"https://open.spotify.com/playlist/{playlist.Id}";
-        Title = playlist.Name ?? "Spotify Playlist";
-    }
-    
-    public string Title { get; }
-    public string Url { get; }
-    public IEnumerable<ITrack> Tracks { get; }
-}
