+++ src/TobysBot.Music/YouTube/YouTubeTrack.cs
@@ -1,23 +1,22 @@
 using Discord;
 using YoutubeExplode.Videos;
 
-namespace TobysBot.Music.Search;
+namespace TobysBot.Music.YouTube;
 
 public class YouTubeTrack : ITrack
 {
     public YouTubeTrack(IVideo video, IUser requestedBy)
     {
-        RequestedBy = requestedBy;
-        Author = video.Author.ChannelTitle;
         Title = video.Title;
-        Url = AudioUrl = video.Url;
-        Duration = video.Duration ?? throw new ArgumentNullException(nameof(video), "Video's duration was null.");
+        Author = video.Author.ChannelTitle;
+        Url = video.Url;
+        Duration = video.Duration ?? throw new ArgumentNullException(nameof(video), "Video duration was null.");
+        RequestedBy = requestedBy;
     }
     
     public string Title { get; }
     public string Author { get; }
     public string Url { get; }
-    public string AudioUrl { get; }
     public TimeSpan Duration { get; }
     public IUser RequestedBy { get; }
 }
